requirements.txt
py-cord==1.7.3
discord.py==1.7.3